build: experimental support for building with Bazel
About two months ago I tried to use Bazel for building our project, but never ended up finishing it.
I still would like to do that, but having this on a separate branch brings a lot of troubles when rebasing,
so I merge this here, but without changing the current GN build system.