feat(backend/config): add support for config file for mintterd
This commit introduces the ability to configure the daemon using a config file in the repo path.
It also removes a "magical" github.com/alecthomas/kong package in favor of a more explicit
configuration for flags and etc. Kong was not able to use a config file in the way we wanted,
and we needed more flexibility for configuration.

Now, an empty config file is always created in the repo path, and we can set any existing configuration flag
in there to override the defaults. Package github.com/peterbourgon/ff is used to combine configuration from
flags, env vars, and config file.