fix(frontend): Write button works on the first screen
for some reason the window rust code does not open the new window with the new draft URL when the current URL is the root (`tauri://localhost`). it does work with other routes.

This removes the route match to the root URL in order to avoid having this error. but this is something that needs to be addressed in the Rust code.