Implement missing Group APIs
This PR implement missing group APIs:

- ListDocumentGroups: to search groups a document is published to.
- ListAccountGroups: to search groups an account is member of.

The change is massive, because it also includes a breaking change to make document IDs unforgeable,
so this doesn't include a migration, but instead marks a new beginning in the list of migrations.
I.e. your current data directory needs to be removed.

It also bumps the network protocol version, to prevent being able to talk to older peers, just to avoid any weird issues.

Frontend Protobuf codegen package changed its name, so all the generated proto files for JS are regenerated.