fix(links): fix click on selection + link
this is an upstream error on tiptap: https://github.com/ueberdosis/tiptap/pull/4346/files

This code will check whether a clicked node name is 'A'. This will fix the issue when openOnClick is enabled, that causes clicking on a selection to open a link that's somewhere within the selection, even though user clicked some other element.